AvoidAlias: deprecate this rule
This rule is detrimental to the powershell community as it undermines
foundational idiomatic syntax that enables powershell to be extremely
powerful with simplicity in mind.

It hinders growth in many projects and forces unnecessary
verbosity around widely used simplistic cmdlets.

Microsoft has provided over 150 built-in aliases that are part of the
standard language and their usage should be allowed. If the problem
is conflicts with aliases then they should be worked out in the
powershell project and not here.
